refactor: client, server error code 분리 #181
