+++ backend/src/main/java/ddangkong/exception/balance/content/NotEnoughBalanceContentException.java
@@ -1,6 +1,6 @@
 package ddangkong.exception.balance.content;
 
-import static ddangkong.exception.ErrorCode.NOT_ENOUGH_BALANCE_CONTENT;
+import static ddangkong.exception.ServerErrorCode.NOT_ENOUGH_BALANCE_CONTENT;
 
 import ddangkong.domain.balance.content.Category;
 import ddangkong.exception.InternalServerException;
