+++ backend/src/main/java/ddangkong/exception/room/NotFoundRoomException.java
@@ -1,6 +1,6 @@
 package ddangkong.exception.room;
 
-import static ddangkong.exception.ErrorCode.NOT_FOUND_ROOM;
+import static ddangkong.exception.ClientErrorCode.NOT_FOUND_ROOM;
 
 import ddangkong.exception.BadRequestException;
