+++ backend/src/main/java/ddangkong/exception/room/balance/roomcontent/MismatchRoundException.java
@@ -1,6 +1,6 @@
 package ddangkong.exception.room.balance.roomcontent;
 
-import static ddangkong.exception.ErrorCode.MISMATCH_ROUND;
+import static ddangkong.exception.ClientErrorCode.MISMATCH_ROUND;
 
 import ddangkong.exception.BadRequestException;
